local/requirements.txt
fastapi==0.104.1           
uvicorn==0.24.0           
gunicorn==21.2.0           
pydantic==2.5.0            
requests==2.32.3           
openai==1.30.0             
PyMuPDF==1.23.8            
azure-storage-blob==12.19.0 
python-multipart==0.0.6    
httpx==0.27.0             

